Decompose components, systems and the engine into the different folders
