+++ src/components/index.ts
@@ -1,0 +1,2 @@
+export {PositionComponent} from './position.component'
+export {VelocityComponent} from './velocity.component'
