+++ src/components/position.component.ts
@@ -1,0 +1,6 @@
+import {Component} from 'typed-ecstasy'
+
+export class PositionComponent extends Component {
+	public x = 0;
+	public y = 0;
+}
