+++ src/systems/index.ts
@@ -1,0 +1,2 @@
+export {RenderSystem} from './render.system'
+export {MovementSystem} from './movement.system'
